requirements.txt
pact-python==1.5.2
fastapi==0.67.0
pytest==5.4.1
requests>=2.26.0
uvicorn>=0.14.0
testcontainers==3.3.0